Finish main function app working
